+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 Unreleased
 ====
 
+0.5.0
+=====
+
+* [#125](https://github.com/kowainik/relude/issues/125):
+  Moved many numerical functions and types in `Relude.Numeric`.    
+  It reexports `toIntegralSized` and exports `integerToBounded` and `integerToNatural`
+
 0.4.0 — Nov 6, 2018
 =====